move some tor things to a tor mod
