+++ crates/ptrs/src/pt_spec_v3.rs
@@ -1,0 +1,3 @@
+//! Pluggable Transport Specification 3.0 API
+//!
+//! TODO: this is in the ideation phase, not a guaranteed feature
